+++ Packages/SSHManagerUI/Sources/SSHManagerUI/Model/AppLinks.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-enum AppLinks {
-    static let repository = "https://github.com/CodeLieutenant/sshconfigmanager"
-    static let newIssue = "https://github.com/CodeLieutenant/sshconfigmanager/issues/new"
-    static let issues = "https://github.com/CodeLieutenant/sshconfigmanager/issues"
-}